change Makefile targets to real files to avoid unneccessary compilation
A lot of Makefile targets are "virtual" files like pred.pdf, add.pdf
that are never generated by any of the scripts. Because a single input
X.asy generates multiple images X000.pdf, X001.pdf, X002.pdf, … , but not
X.pdf. This causes each make run to generate *every* image again *every
time*, even if they are still up-to-date.
This change uses the first of each generated images, X000.pdf, as targets.

Note: technically *all* generated files should be used as targets, but
this is good enough for now and already used in the main Makefile.